+++ packages/react-static-plugin-reach-router/src/browser.api.js
@@ -2,25 +2,40 @@ import React from 'react'
 import { useBasepath, useStaticInfo, makePathAbsolute } from 'react-static'
 import { Router, ServerLocation } from '@reach/router'
 
-const DefaultPath = ({ render }) => render
-
 export default ({ RouterProps: userRouterProps = {} }) => ({
-  Root: PreviousRoot => ({ children }) => {
+  Root: PreviousRoot => ({ children, ...rest }) => {
+    const Render = ({ render, ...rest }) => render(rest)
     const basepath = useBasepath()
     const staticInfo = useStaticInfo()
 
-    children = (
-      <Router {...userRouterProps} basepath={basepath}>
-        <DefaultPath default render={<PreviousRoot>{children}</PreviousRoot>} />
+    const renderedChildren = (
+      // Place a top-level router inside the root
+      // This will give proper context to Link and other reach components
+      <Router {...(basepath ? { basepath } : {})} {...userRouterProps}>
+        <Render
+          path="/*"
+          render={location =>
+            console.log({ location }) || (
+              <PreviousRoot {...rest}>{children}</PreviousRoot>
+            )
+          }
+        />
       </Router>
     )
 
+    // If we're in SSR, use a manual server location
     return typeof document === 'undefined' ? (
       <ServerLocation url={makePathAbsolute(staticInfo.path)}>
-        {children}
+        {renderedChildren}
       </ServerLocation>
     ) : (
-      children
+      renderedChildren
     )
   },
+  Routes: PreviousRoutes => props => (
+    // Wrap Routes in a reach router
+    <Router>
+      <PreviousRoutes path="/*" {...props} />} />
+    </Router>
+  ),
 })