+++ packages/react-static/src/browser/components/Root.js
@@ -2,27 +2,15 @@ import React from 'react'
 //
 import { plugins } from '../'
 import { makeHookReducer } from '../utils'
-import ErrorBoundary from './ErrorBoundary'
-import { useStaticInfo } from '../hooks/useStaticInfo'
-import { useBasepath } from '../hooks/useBasepath'
-import Spinner from './Spinner'
-
-const RootWrapper = ({ children }) => children
-
-const RootHook = makeHookReducer(plugins, 'Root', { sync: true })
-const ResolvedRoot = RootHook(RootWrapper)
 
 export const Root = ({ children }) => {
-  const staticInfo = useStaticInfo()
-  const basepath = useBasepath()
-
-  return (
-    <ErrorBoundary>
-      <React.Suspense fallback={<Spinner />}>
-        <ResolvedRoot basepath={basepath} staticInfo={staticInfo}>
-          {children}
-        </ResolvedRoot>
-      </React.Suspense>
-    </ErrorBoundary>
+  const ResolvedRoot = React.useMemo(
+    () => {
+      const RootHook = makeHookReducer(plugins, 'Root', { sync: true })
+      return RootHook(({ children }) => children)
+    },
+    [plugins]
   )
+
+  return <ResolvedRoot>{children}</ResolvedRoot>
 }