+++ packages/react-static/src/browser/hooks/useRouteData.js
@@ -13,13 +13,18 @@ export const useRouteData = () => {
   // unless there is data for the 404 page
   if (routeError && (!routeInfo || !routeInfo.data)) {
     throw new Error(
-      `React-Static: <RouteData> could not find any data for this route: ${routePath}. If this is a dynamic route, please remove any reliance on RouteData or withRouteData from this routes components`
+      `React-Static: useRouteData() could not find any data for this route: ${routePath}. If this is a dynamic route, please remove any calls to useRouteData() from this route's components`
     )
   }
 
   // If we need to load data, suspend while it's requested
   if (shouldLoadData(routeInfo)) {
-    throw prefetch(routePath, { priority: true })
+    throw Promise.all([
+      new Promise(resolve =>
+        setTimeout(resolve, process.env.REACT_STATIC_MIN_LOAD_TIME)
+      ),
+      prefetch(routePath, { priority: true }),
+    ])
   }
 
   // Otherwise, return all of the data