+++ packages/react-static/src/index.js
@@ -4,6 +4,7 @@ export { Helmet as Head } from 'react-helmet'
 // React-Static
 export { Root } from './browser/components/Root'
 export { Routes } from './browser/components/Routes'
+export { default as ErrorBoundary } from './browser/components/ErrorBoundary'
 export { useRouteData } from './browser/hooks/useRouteData'
 export { useSiteData } from './browser/hooks/useSiteData'
 export { usePrefetch } from './browser/hooks/usePrefetch'